requirements.txt
python-frontmatter
click
mistune
python-slugify
tomli
yoix-pi
pybars3